Added models and related functions into separate file
